spin_locked_app

Module locked

Spin lock file (spin.lock) serialization models.

Structs

ContentPath

A ContentPath specifies content mapped to a WASI path.

ContentRef

A ContentRef represents content used by an application.

LockedApp

A LockedApp represents a “fully resolved” Spin application.

LockedComponent

A LockedComponent represents a “fully resolved” Spin component.

LockedComponentDependency

A LockedDependency represents a “fully resolved” Spin component dependency.

LockedComponentSource

A LockedComponentSource specifies a Wasm source.

LockedTrigger

A LockedTrigger specifies configuration for an application trigger.

Variable

A Variable specifies a custom configuration variable.

Enums

InheritConfiguration

InheritConfiguration specifies which configurations to inherit from parent.

MustUnderstand

Identifies fields in the LockedApp that the host must process if present.

Constants

HOST_REQ_OPTIONAL

Indicates that a host feature is optional. This is the default and is equivalent to omitting the feature from host_requirements.

HOST_REQ_REQUIRED

Indicates that a host feature is required.

SERVICE_CHAINING_KEY

If present and required in host_requirements, the host must support local service chaining (*.spin.internal) or reject the app.

Type Aliases

LockedMap

A String-keyed map with deterministic serialization order.